% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{AlgorithmOptions}
\alias{AlgorithmOptions}
\alias{AlgorithmOptions_HJKB}
\alias{AlgorithmOptions_BOBYQA}
\alias{AlgorithmOptions_DEoptim}
\title{Algorithm Options for Optimization Algorithms}
\format{
An object of class \code{list} of length 5.

An object of class \code{list} of length 6.

An object of class \code{list} of length 21.
}
\usage{
AlgorithmOptions_HJKB

AlgorithmOptions_BOBYQA

AlgorithmOptions_DEoptim
}
\description{
Default options for various optimization algorithms used within the package.
These algorithms are configured via the \code{PIConfiguration} class.
}
\section{AlgorithmOptions_HJKB}{
 Default options for the \code{HJKB} algorithm. See
\code{\link[dfoptim:hookejeeves]{dfoptim::hjkb()}} for more details on the options.
}

\section{AlgorithmOptions_BOBYQA}{
 Default options for the \code{BOBYQA} algorithm.
See \code{\link[nloptr:nl.opts]{nloptr::nl.opts()}} for more details on the options.
}

\section{AlgorithmOptions_DEoptim}{
 Default options for the \code{DEoptim}
algorithm. See \code{\link[DEoptim:DEoptim.control]{DEoptim::DEoptim.control()}} for more details on the
options.
}

\keyword{datasets}
