% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{AlgorithmOptions_BOBYQA}
\alias{AlgorithmOptions_BOBYQA}
\title{AlgorithmOptions_BOBYQA}
\format{
An object of class \code{list} of length 7.
}
\usage{
AlgorithmOptions_BOBYQA
}
\description{
Default options for the BOBYQA algorithm.
see \code{\link[nloptr:nl.opts]{nloptr::nl.opts()}} for details.
}
\keyword{datasets}
