% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{AlgorithmOptions_DEoptim}
\alias{AlgorithmOptions_DEoptim}
\title{AlgorithmOptions_DEoptim}
\format{
An object of class \code{list} of length 21.
}
\usage{
AlgorithmOptions_DEoptim
}
\description{
Default options for the DEoptim algorithm.
see \code{\link[DEoptim:DEoptim.control]{DEoptim::DEoptim.control()}} for details.
}
\keyword{datasets}
