% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{AlgotitmOptions_HJKB}
\alias{AlgotitmOptions_HJKB}
\title{AlgotitmOptions_HJKB}
\format{
An object of class \code{list} of length 5.
}
\usage{
AlgotitmOptions_HJKB
}
\description{
Default options for the HJKB algorithm.
see \code{\link[dfoptim:hookejeeves]{dfoptim::hjkb()}} for details.
}
\keyword{datasets}
