% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{.getMolWeightFor}
\alias{.getMolWeightFor}
\title{Retrieve molecular weight for the molecule of a quantity}
\usage{
.getMolWeightFor(quantity)
}
\arguments{
\item{quantity}{A \code{Quantity} object.}
}
\value{
Numeric value of the molecular weight in \code{kg/µmol}.
}
\description{
Gets the molecular weight of a quantity's parent molecule in
\code{kg/µmol}.
}
\keyword{internal}
