% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{.getParentContainer}
\alias{.getParentContainer}
\title{Get the parent container of the entity}
\usage{
.getParentContainer(entity, type)
}
\arguments{
\item{entity}{Object of type \code{Entity}}
}
\value{
The container type that is the parent of the entity.
}
\description{
Get the parent container of the entity
}
\keyword{internal}
