% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-object-functions.R
\name{.handleSimulationFailure}
\alias{.handleSimulationFailure}
\title{Handle Simulation Failure}
\usage{
.handleSimulationFailure(simulationResults)
}
\arguments{
\item{simulationResults}{A list of simulation results.}
}
\value{
Returns an infinite cost structure if NA values are found in the simulation results, NULL otherwise.
}
\description{
This function checks for NA in simulation results and returns an infinite cost structure if any are found.
}
\keyword{internal}
