% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{.isSameDimension}
\alias{.isSameDimension}
\title{Do all quantities have the same dimension?}
\usage{
.isSameDimension(quantities)
}
\arguments{
\item{quantities}{Quantities which dimensions are compared}
}
\value{
TRUE if the dimension of all quantities are equal, otherwise FALSE
}
\description{
Do all quantities have the same dimension?
}
\keyword{internal}
