% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{.removeFormulaPaths}
\alias{.removeFormulaPaths}
\title{Remove paths with formulas}
\usage{
.removeFormulaPaths(paths, simulation, stopIfNotFound = TRUE)
}
\arguments{
\item{paths}{List of paths to be filtered}

\item{simulation}{A \code{Simulation} object containing the quantities}

\item{stopIfNotFound}{Boolean. If \code{TRUE} (default), an error is thrown when a
path is not found in the simulation.}
}
\value{
List of quantity paths that are not defined by explicit formula.
}
\description{
Removes paths to quantities that are defined by an explicit
formula in the simulation
}
\keyword{internal}
