% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{.restoreSimulationState}
\alias{.restoreSimulationState}
\title{Restore simulation output state}
\usage{
.restoreSimulationState(simulations, simStateList)
}
\arguments{
\item{simulations}{List of \code{Simulation} objects}

\item{simStateList}{Output of the function \code{.storeSimulationState}.}
}
\description{
Restore simulation output state
}
\details{
A simulation whose output state has not changed is left as it is,
so that the output interval objects of its output schema stay the same.
}
\keyword{internal}
