% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-simulation.R
\name{.storeSimulationState}
\alias{.storeSimulationState}
\title{Stores current simulation output state}
\usage{
.storeSimulationState(simulations)
}
\arguments{
\item{simulations}{List of \code{Simulation} objects}
}
\value{
A named list with names being the IDs of the simulations. Every
entry is a list with the paths of the output selections (\code{outputs}), the
output time points (\code{timePoints}), and the name, start time, end time and
resolution of every output interval (\code{intervals}).
}
\description{
Stores the output selections, output intervals, and output
time points of simulations in their current state.
}
\keyword{internal}
