% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-objective-function.R
\name{.summarizeCostLists}
\alias{.summarizeCostLists}
\title{Summarize Cost Lists}
\usage{
.summarizeCostLists(list1, list2)
}
\arguments{
\item{list1}{The first list, containing the output of the
\code{.calculateCostMetrics} function, which includes \code{modelCost},
\code{minLogProbability}, \code{objectiveType}, \code{costVariables}, and
\code{residualDetails}.}

\item{list2}{The second list, containing the output of the
\code{.calculateCostMetrics} function, which includes \code{modelCost},
\code{minLogProbability}, \code{objectiveType}, \code{costVariables}, and
\code{residualDetails}.}
}
\value{
Returns a list that includes the sum of \code{modelCosts}, the sum of
\code{minLogProbabilities}, the \code{objectiveType} taken from \code{list1}, a row-bound
combination of \code{costVariables}, and a row-bound combination of
\code{residualDetails}.
}
\description{
This function takes two lists, each being the output of the
\code{.calculateCostMetrics} function, and summarizes them. It aggregates model
costs and min log probabilities, and combines cost and residual details by
row-binding.
}
\keyword{internal}
