% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{.validateIsSameDimension}
\alias{.validateIsSameDimension}
\title{Check if the dimensions of the quantities are equal.}
\usage{
.validateIsSameDimension(quantities)
}
\arguments{
\item{quantities}{Quantities which dimensions are compared}
}
\description{
Check if the dimensions of the quantities are equal.
}
\keyword{internal}
