% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{.validateOutputMappingHasData}
\alias{.validateOutputMappingHasData}
\title{Validate Observed Data Availability in PIOutputMapping}
\usage{
.validateOutputMappingHasData(object)
}
\arguments{
\item{object}{A PIOutputMapping object or a list of PIOutputMapping objects.}
}
\description{
Ensures each PIOutputMapping object has observed datasets. Throws an error
if no observed data is found.
}
\keyword{internal}
