% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuitepi-env.R
\name{getOSPSuitePISetting}
\alias{getOSPSuitePISetting}
\title{Get the value of a global ospsuite.parameteridentification setting.}
\usage{
getOSPSuitePISetting(settingName)
}
\arguments{
\item{settingName}{String name of the setting}
}
\value{
Value of the setting stored in \code{piEnv}. If the setting does not exist,
an error is thrown.
}
\description{
Get the value of a global ospsuite.parameteridentification setting.
}
\examples{
getOSPSuitePISetting("packageVersion")
}
