% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ObjectiveFunctions}
\alias{ObjectiveFunctions}
\title{ObjectiveFunctions
List of supported objective functions to calculate the error.}
\format{
An object of class \code{list} of length 2.
}
\usage{
ObjectiveFunctions
}
\description{
ObjectiveFunctions
List of supported objective functions to calculate the error.
}
\keyword{datasets}
