% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ObjectiveFunctionSpecs}
\alias{ObjectiveFunctionSpecs}
\title{Objective Function Specifications}
\format{
An object of class \code{list} of length 7.
}
\usage{
ObjectiveFunctionSpecs
}
\description{
Specifies supported objective function configurations for error calculation
in parameter optimization. These specifications detail the allowable options
for tailoring how model fit is assessed, configured within \code{PIConfiguration}.
}
\details{
Includes:
\itemize{
\item \strong{\code{objectiveFunctionType}} – Type of error calculation. Allowed values:
\code{lsq} (least squares), \verb{m3"} (M3 method).
\item \strong{\code{residualWeightingMethod}} – Method for weighting residuals. Allowed values:
\code{none}, \code{std}, \code{mean}, \code{error}.
\item \strong{\code{robustMethod}} – Approach for robust outlier handling. Allowed values:
\code{none}, \code{huber}, \code{bisquare}.
\item \strong{\code{scaleVar}} – Boolean for variance scaling. Allowed values: \code{TRUE}, \code{FALSE}.
\item \strong{\code{scaling}} – Scaling type. Allowed values: \code{lin} (linear), \code{log} (logarithmic).
\item \strong{\code{linScaleCV}} – Coefficient of variation for linear scaling. Numeric
range: \code{1e-9} to \code{1}.
\item \strong{\code{logScaleSD}} – Standard deviation for log scaling. Numeric range: \code{1e-9}
to \code{Inf}.
}

These options directly influence the optimization process by defining how
discrepancies between simulated and observed data are quantified and managed.
}
\keyword{datasets}
