% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuitepi-env.R
\name{ospsuitePIRSettingNames}
\alias{ospsuitePIRSettingNames}
\title{Names of the settings stored in piEnv Can be used with
\code{getOSPSuitePISetting()}}
\usage{
ospsuitePIRSettingNames
}
\description{
Names of the settings stored in piEnv Can be used with
\code{getOSPSuitePISetting()}
}
