% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-object-functions.R
\name{plot.modelCost}
\alias{plot.modelCost}
\title{Plot Model Cost Residuals}
\usage{
\method{plot}{modelCost}(x, legpos = "topright", ...)
}
\arguments{
\item{x}{A \code{modelCost} object containing residuals to plot.}

\item{legpos}{Position of the legend; default is "topright". Use NA to omit the legend.}

\item{...}{Additional arguments passed to the plot function.}
}
\value{
Generates a plot.
}
\description{
Plots raw residuals and, if different, weighted and robust weighted residuals
from a \code{modelCost} object.
}
\examples{
# Assuming modelCostObj is a valid `modelCost` object
\dontrun{
plot.modelCost(modelCostObj)
}
}
