% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-results.R
\name{print.piResult}
\alias{print.piResult}
\title{Print Parameter Identification Results}
\usage{
\method{print}{piResult}(x)
}
\arguments{
\item{x}{An S3 object of class \code{piResult}.}
}
\value{
The input \code{piResult} object, invisibly.
}
\description{
Custom \code{print()} method for \code{piResult} objects. Displays a
summary including algorithm, convergence status, objective value, estimated
parameters and confidence intervals.
}
