% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{robustMethodOptions}
\alias{robustMethodOptions}
\title{Robust Weighting Methods for Cost Function}
\usage{
robustMethodOptions
}
\description{
Robust weighting methods to address outliers in the cost function calculation
during parameter optimization.
}
\details{
The available methods are:
\itemize{
\item \strong{\code{none}} - No robust weighting applied.
\item \strong{\code{huber}} - Huber weighting for moderate outliers.
\item \strong{\code{bisquare}} - Bisquare (Tukey's biweight) weighting for severe outliers.
}
}
