% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-helper.R
\name{validateIsOption}
\alias{validateIsOption}
\title{Validate Input Options Against Valid Options}
\usage{
validateIsOption(inputOptions, validOptions)
}
\arguments{
\item{inputOptions}{A list of options to validate. Each option's value is checked against
the corresponding entry in \code{validOptions}.}

\item{validOptions}{A list that specifies valid values or value types for each option.
It can include specific values, types, or ranges for numeric options.}
}
\value{
\itemize{
\item If all input options are valid, the function returns \code{NULL}.
\item If any input option is invalid, the function stops and signals an error with a
message detailing the validation failures.
}
}
\description{
This function checks if the provided input options are valid based on a list of valid options.
It ensures that each option in \code{inputOptions} matches the allowed types or values specified
in \code{validOptions}. For options expecting numeric values, it checks if they fall within a
specified range. For boolean options, it checks if they are strictly \code{TRUE} or \code{FALSE}.
}
\examples{
validOptions <- list(
  algorithmType = c("gradient", "newton"),
  enableLogging = c(TRUE, FALSE),
  convergenceThreshold = list(type = "numeric", min = 1e-5, max = 0.1)
)

inputOptions <- list(
  algorithmType = "gradient",
  enableLogging = TRUE,
  convergenceThreshold = 0.005
)
validateIsOption(inputOptions, validOptions)

}
