% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{AlgorithmOptions}
\alias{AlgorithmOptions}
\title{AlgorithmOptions
List of options for optimization algorithms, some of them specific certain algorithms.
See documentation of \code{FME::modeFit()} for details.}
\format{
An object of class \code{list} of length 48.
}
\usage{
AlgorithmOptions
}
\description{
AlgorithmOptions
List of options for optimization algorithms, some of them specific certain algorithms.
See documentation of \code{FME::modeFit()} for details.
}
\keyword{datasets}
