% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{Algorithms}
\alias{Algorithms}
\title{Algorithms
List of optimization algorithms supported by optimization routines and, consequently,
by the ParameterIdentification class}
\format{
An object of class \code{list} of length 18.
}
\usage{
Algorithms
}
\description{
Algorithms
List of optimization algorithms supported by optimization routines and, consequently,
by the ParameterIdentification class
}
\keyword{datasets}
