% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{Algorithms}
\alias{Algorithms}
\title{Optimization Algorithms}
\format{
An object of class \code{list} of length 3.
}
\usage{
Algorithms
}
\description{
Optimization algorithms supported by optimization routines, available for use
in the \code{ParameterIdentification} class. These algorithms are configured via
the \code{PIConfiguration} class.
}
\details{
Supported algorithms include:
\itemize{
\item \strong{\code{HJKB}} – Hooke-Jeeves algorithm from the \pkg{dfoptim} package.
\item \strong{\code{BOBYQA}} – BOBYQA algorithm from the \pkg{nloptr} package.
\item \strong{\code{DEoptim}} – Differential evolution algorithm from the \pkg{DEoptim}
package, suitable for stochastic global optimization.
}

These algorithms can be specified and configured within the \code{PIConfiguration}
class to tailor the parameter identification process to specific needs.
}
\keyword{datasets}
