% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-object-functions.R
\name{.applyLogTransformation}
\alias{.applyLogTransformation}
\title{Apply Log Transformation to Data Frame}
\usage{
.applyLogTransformation(df)
}
\arguments{
\item{df}{A \code{tbl_df} representing the observed vs predicted data frame (\code{obsVsPredDf}).}
}
\value{
A transformed data frame with log-transformed \code{yValues} and \code{lloq}.
}
\description{
Transforms the \code{yValues} and \code{lloq} columns in the given data frame using a log
transformation. Currently, this function only supports \code{obsVsPredDf} data frames,
which must contain \code{yDimension}, \code{yUnit}, \code{yValues}, and \code{lloq} columns.
}
\examples{
# Assuming df is a valid obsVsPredDf data frame
\dontrun{
transformedDf <- applyLogTransformation(df)
}
}
\keyword{internal}
