% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-object-functions.R
\name{.createErrorCostStructure}
\alias{.createErrorCostStructure}
\title{Create Infinite Cost Structure}
\usage{
.createErrorCostStructure(infinite = FALSE)
}
\value{
Returns a list with infinite values for model, minlogp, var, and residuals components.
}
\description{
Generates a cost structure with infinite values, used in cases of simulation failure.
}
\keyword{internal}
