% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-quantity.R
\name{.validateOutputMappingHasData}
\alias{.validateOutputMappingHasData}
\title{Validate observed data availability in \code{PIOutputMapping}}
\usage{
.validateOutputMappingHasData(object)
}
\arguments{
\item{object}{A \code{PIOutputMapping} object or a list of \code{PIOutputMapping}
objects.}
}
\description{
Ensures each \code{PIOutputMapping} object has observed datasets. Throws an error
if no observed data is found.
}
\keyword{internal}
