% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ObjectiveFunctionOptions}
\alias{ObjectiveFunctionOptions}
\title{ObjectiveFunctionOptions
List of supported parameter options for calculation of model cost.}
\format{
An object of class \code{list} of length 7.
}
\usage{
ObjectiveFunctionOptions
}
\description{
ObjectiveFunctionOptions
List of supported parameter options for calculation of model cost.
}
\keyword{datasets}
