% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ObjectiveFunctionOptions}
\alias{ObjectiveFunctionOptions}
\title{ObjectiveFunctions
List of supported objective functions to calculate the error.}
\format{
An object of class \code{list} of length 7.
}
\usage{
ObjectiveFunctionOptions
}
\description{
ObjectiveFunctions
List of supported objective functions to calculate the error.
}
\keyword{datasets}
