% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ObjectiveFunctionSpecs}
\alias{ObjectiveFunctionSpecs}
\title{Objective Function Specifications}
\format{
An object of class \code{list} of length 7.
}
\usage{
ObjectiveFunctionSpecs
}
\description{
Specifies supported objective function configurations for error calculation
in parameter optimization. These specifications detail the allowable options
for tailoring how model fit is assessed, configured within \code{PIConfiguration}.
}
\details{
Includes:
\itemize{
\item \code{objectiveFunctionType}: Types of error calculations ("lsq" for least squares, "m3" for method 3).
\item \code{residualWeightingMethod}: Methods for weighting residuals ("none", "std", "mean", "error").
\item \code{robustMethod}: Approaches for robust outlier handling ("none", "huber", "bisquare").
\item \code{scaleVar}: Boolean for variance scaling (TRUE, FALSE).
\item \code{linScaleCV}: Coefficient of variation for linear scale, with numeric range 1e-9 to 1.
\item \code{logScaleSD}: Standard deviation for log scale, with numeric range 1e-9 to Inf.
}
These options directly influence the optimization process by defining how
discrepancies between simulated and observed data are quantified and managed.
}
\keyword{datasets}
