% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuitepi-env.R
\docType{data}
\name{ospsuitePIRSettingNames}
\alias{ospsuitePIRSettingNames}
\title{Names of the settings stored in piEnv Can be used with
\code{getOSPSuitePISetting()}}
\format{
An object of class \code{list} of length 2.
}
\usage{
ospsuitePIRSettingNames
}
\description{
Names of the settings stored in piEnv Can be used with
\code{getOSPSuitePISetting()}
}
\keyword{datasets}
