% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterIdentification.R
\docType{class}
\name{ParameterIdentification}
\alias{ParameterIdentification}
\title{ParameterIdentification}
\description{
A task to identify optimal parameter values based on simulation
outputs and observed data
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{ParameterIdentification}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{simulations}}{Named list with simulation objects, where names are
IDs of the root container of the simulation}

\item{\code{parameters}}{List of \code{PIParameters} objects to be optimized. Read-only}

\item{\code{configuration}}{An object of \code{PIConfiguration}}

\item{\code{outputMappings}}{List of \code{PIOutputMapping} objects. Each
mapping assigns a set of observed data to a simulation output
Evaluate all simulations with given parameter values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParameterIdentification-new}{\code{ParameterIdentification$new()}}
\item \href{#method-ParameterIdentification-run}{\code{ParameterIdentification$run()}}
\item \href{#method-ParameterIdentification-plotResults}{\code{ParameterIdentification$plotResults()}}
\item \href{#method-ParameterIdentification-gridSearch}{\code{ParameterIdentification$gridSearch()}}
\item \href{#method-ParameterIdentification-calculateOFVProfiles}{\code{ParameterIdentification$calculateOFVProfiles()}}
\item \href{#method-ParameterIdentification-print}{\code{ParameterIdentification$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$new(
  simulations,
  parameters,
  outputMappings,
  configuration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulations}}{An object or a list of objects of class \code{Simulation}.
Parameters of the simulation object will be varied and the results simulated}

\item{\code{parameters}}{An object or a list of objects of class \code{PIParameter}. These parameters will be varied.}

\item{\code{outputMappings}}{List of objects of the class \code{PIOutputMapping}. Each objects
maps a model output (represented by a \code{Quantity}) with a set of observed data given as \code{XYData} objects.
is used.}

\item{\code{configuration}}{Optional. Object of type \code{PIConfiguration} defining
further options of the parameter identification. If no \code{PIConfiguration} is passed, a default one}

\item{\code{currVals}}{Numerical vector of the parameter values to be applied}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An list of objects of \code{DataCombined} class that includes values simulated
with the given parameters, and corresponding datasets with observed data.
Returns one \code{DataCombined} object for each output mapping.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-run"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-run}{}}}
\subsection{Method \code{run()}}{
Start identification of parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$run()}\if{html}{\out{</div>}}
}

\subsection{Details}{
When the identification if finished, the best identified values of the parameters
are accessible via the \code{currValue}-field of the \code{PIParameters}-object.
}

\subsection{Returns}{
Output of the PI algorithm. Depends on the selected algorithm.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-plotResults"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-plotResults}{}}}
\subsection{Method \code{plotResults()}}{
Plot the results of parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$plotResults(par = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{Values of paramterers to be applied to the simulations.
If \code{NULL} (default), current parameter values are applied. If custom
values are supplied, the they must be in the same order as \code{ParameterIdentification$parameters}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Runs all simulations with current (default) or supplied
parameter values and creates plots of every output mapping
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-gridSearch"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-gridSearch}{}}}
\subsection{Method \code{gridSearch()}}{
Calculates the values of the objective function on an n-dimensional grid, where n is the number
of parameters, and optionally saves the best result as the starting point for next optimization runs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$gridSearch(
  lower = NULL,
  upper = NULL,
  logScaleFlag = FALSE,
  totalEvaluations = 50,
  margin = 0,
  setStartingPoint = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lower}}{A vector of lower bounds for parameters, with the same length as the number of parameters
optimized in this parameter identification task. By default, uses the minimal values
defined in the \code{PIParameter} objects.}

\item{\code{upper}}{A vector of upper bounds for parameters, with the same length as the number of parameters
optimized in this parameter identification task. By default, uses the maximal values
defined in the \code{PIParameter} objects.}

\item{\code{logScaleFlag}}{A single logical value or a vector of logical values indicating
if grid should be evenly spaced on a linear or a logarithmic scale. Defaults to \code{FALSE}.}

\item{\code{totalEvaluations}}{An integer number. The grid will have as many points so that the
total number of grid points does not exceed \code{totalEvaluations}. Defaults to \code{50}.}

\item{\code{margin}}{Can be set to a non-zero positive value so that the edges of the grid will be away
from the exact parameter bounds.}

\item{\code{setStartingPoint}}{(logical) If \code{TRUE}, the best result will be saved as the starting point for
the next optimization runs. Defaults to \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble with one column for each parameter and one column for the objective function value.
The tibble will have at most \code{totalEvaluations} rows.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-calculateOFVProfiles"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-calculateOFVProfiles}{}}}
\subsection{Method \code{calculateOFVProfiles()}}{
Calculates the values of the objective function on all orthogonal lines
passing through a given point in the parameter space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$calculateOFVProfiles(
  par = NULL,
  lower = NULL,
  upper = NULL,
  totalEvaluations = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{A vector of parameter values, with the same length as the number of parameters.
If not supplied, the current parameter values are used.}

\item{\code{lower}}{A vector of lower bounds for parameters, with the same length as the number of parameters.
By default, uses 0.9 of the current parameter value.}

\item{\code{upper}}{A vector of upper bounds for parameters, with the same length as the number of parameters.
By default, uses 1.1 of the current parameter value.}

\item{\code{totalEvaluations}}{An integer number. The combined profiles will not contain more than \code{totalEvaluations}
points. If not supplied, 21 points per parameter are plotted to cover a uniform grid from 0.9 to 1.1.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of tibbles, one tibble per parameter, with one column for parameter values
and one column for the matching objective function values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-print"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
