% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterIdentification.R
\docType{class}
\name{ParameterIdentification}
\alias{ParameterIdentification}
\title{ParameterIdentification}
\description{
A task to identify optimal parameter values based on simulation outputs and observed data
}
\section{Super class}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{ParameterIdentification}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{simulations}}{Named list with simulation objects, where names are IDs of the root container of the simulation}

\item{\code{parameters}}{List of \code{PIParameters} objects to be optimized. Read-only}

\item{\code{configuration}}{An object of \code{PIConfiguration}}

\item{\code{outputMappings}}{List of \code{PIOutputMapping} objects. Each mapping assigns a set of observed data
given by \code{XYData}-objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParameterIdentification$new()}}
\item \href{#method-finalize}{\code{ParameterIdentification$finalize()}}
\item \href{#method-run}{\code{ParameterIdentification$run()}}
\item \href{#method-plotCurrentResults}{\code{ParameterIdentification$plotCurrentResults()}}
\item \href{#method-print}{\code{ParameterIdentification$print()}}
\item \href{#method-clone}{\code{ParameterIdentification$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$new(
  simulations,
  parameters,
  outputMappings,
  configuration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulations}}{An object or a list of objects of class \code{Simulation}.
Parameters of the simulation object will be varied and the results simulated}

\item{\code{parameters}}{An object or a list of objects of class \code{PIParameter}. These parameters will be varied.}

\item{\code{outputMappings}}{List of objects of the class \code{PIOutputMapping}. Each objects
maps a model output (represented by a \code{Quantity}) with a set of observed data given as \code{XYData} objects.
is used.}

\item{\code{configuration}}{Optional. Object of type \code{PIConfiguration} defining
further options of the parameter identification. If no \code{PIConfiguration} is passed, a default one}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ParameterIdentification` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clean up upon object removal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Start identification of parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$run()}\if{html}{\out{</div>}}
}

\subsection{Details}{
When the identification if finished, the best identified values of the parameters are accessible via the \code{currValue}-field of the \code{PIParameters}-object.
}

\subsection{Returns}{
Output of the PI algorithm. Depends on the selected algorithm.
Plot the current results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCurrentResults"></a>}}
\if{latex}{\out{\hypertarget{method-plotCurrentResults}{}}}
\subsection{Method \code{plotCurrentResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$plotCurrentResults()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Runs all simulations with current parameter values and creates plots of every output mapping
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
