% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterIdentification.R
\docType{class}
\name{ParameterIdentification}
\alias{ParameterIdentification}
\title{ParameterIdentification}
\description{
A task to identify optimal parameter values based on simulation
outputs and observed data
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{ParameterIdentification}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{simulations}}{Named list with simulation objects, where names are
IDs of the root container of the simulation}

\item{\code{parameters}}{List of \code{PIParameters} objects to be optimized. Read-only}

\item{\code{configuration}}{An object of \code{PIConfiguration}}

\item{\code{outputMappings}}{List of \code{PIOutputMapping} objects. Each
mapping assigns a set of observed data to a simulation output
Evaluate all simulations with given parameter values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParameterIdentification-new}{\code{ParameterIdentification$new()}}
\item \href{#method-ParameterIdentification-run}{\code{ParameterIdentification$run()}}
\item \href{#method-ParameterIdentification-plotResults}{\code{ParameterIdentification$plotResults()}}
\item \href{#method-ParameterIdentification-print}{\code{ParameterIdentification$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$new(
  simulations,
  parameters,
  outputMappings,
  configuration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulations}}{An object or a list of objects of class \code{Simulation}.
Parameters of the simulation object will be varied and the results simulated}

\item{\code{parameters}}{An object or a list of objects of class \code{PIParameter}. These parameters will be varied.}

\item{\code{outputMappings}}{List of objects of the class \code{PIOutputMapping}. Each objects
maps a model output (represented by a \code{Quantity}) with a set of observed data given as \code{XYData} objects.
is used.}

\item{\code{configuration}}{Optional. Object of type \code{PIConfiguration} defining
further options of the parameter identification. If no \code{PIConfiguration} is passed, a default one}

\item{\code{currVals}}{Numerical vector of the parameter values to be applied}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An list of objects of \code{DataCombined} class that includes values simulated
with the given parameters, and corresponding datasets with observed data.
Returns one \code{DataCombined} object for each output mapping.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-run"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-run}{}}}
\subsection{Method \code{run()}}{
Start identification of parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$run()}\if{html}{\out{</div>}}
}

\subsection{Details}{
When the identification if finished, the best identified values of the parameters are accessible via the \code{currValue}-field of the \code{PIParameters}-object.
}

\subsection{Returns}{
Output of the PI algorithm. Depends on the selected algorithm.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-plotResults"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-plotResults}{}}}
\subsection{Method \code{plotResults()}}{
Plot the results of parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$plotResults(par = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{Values of paramterers to be applied to the simulations.
If \code{NULL} (default), current parameter values are applied. If custom
values are supplied, the they must be in the same order as \code{ParameterIdentification$parameters}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Runs all simulations with current (default) or supplied
parameter values and creates plots of every output mapping
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParameterIdentification-print"></a>}}
\if{latex}{\out{\hypertarget{method-ParameterIdentification-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterIdentification$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
