% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIConfiguration.R
\docType{class}
\name{PIConfiguration}
\alias{PIConfiguration}
\title{PIConfiguration}
\description{
An object storing configuration for the parameter identification
}
\section{Super class}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{PIConfiguration}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{simulateSteadyState}}{Boolean representing whether the simulation will be brought to a steady-state first}

\item{\code{steadyStateTime}}{Time in minutes to simulate if simulating steady-state. May be NULL}

\item{\code{printIterationFeedback}}{Boolean. If TRUE, the output of the residuals calculation will be printed after each iteration.
Default is FALSE}

\item{\code{numberOfCores}}{(Maximal) number of cores to be used. This is only relevant when \code{parallellize}
is \code{TRUE}.
Default is getOSPSuitePISetting("maxNumberOfCores").}

\item{\code{parallelize}}{Logical. If \code{TRUE} (default), simulations for each evaluation are executed in parallel.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PIConfiguration$new()}}
\item \href{#method-print}{\code{PIConfiguration$print()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `PIConfiguration` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
}
