% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIConfiguration.R
\docType{class}
\name{PIConfiguration}
\alias{PIConfiguration}
\title{PIConfiguration}
\description{
Encapsulates configurations such as optimization algorithm
choice, and evaluation settings for parameter identification.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{printEvaluationFeedback}}{Boolean. If \code{TRUE}, prints objective
function value after each evaluation. Default is \code{FALSE}.}

\item{\code{simulationRunOptions}}{Object of \code{SimulationRunOptions} for
simulation runs. If \code{NULL}, default options are used.}

\item{\code{objectiveFunctionOptions}}{Settings for model fit evaluation,
affecting error metrics and cost calculation. See
\code{\link{ObjectiveFunctionSpecs}} for details. Defaults in
\code{\link{ObjectiveFunctionOptions}}.}

\item{\code{algorithm}}{Optimization algorithm name. See
\code{\link{Algorithms}} for a list of
supported algorithms. Defaults to \code{BOBYQA}.}

\item{\code{ciMethod}}{Confidence interval estimation method. See
\code{\link{CIMethods}} for available options.
Defaults to \code{hessian}.}

\item{\code{algorithmOptions}}{Named list of settings specific to the selected
algorithm.. Refer to
\code{\link{AlgorithmOptions}} for default
settings per algorithm (e.g., \code{AlgorithmOptions_XYZ} where \code{XYZ}
denotes the algorithm name). If \code{NULL}, algorithm's default settings
are applied.}

\item{\code{ciOptions}}{Named list of settings for the selected CI method. Refer
to \code{\link{CIOptions}} for default settings
per method (e.g., \code{CIOptions_XYZ} where \code{XYZ} corresponds to the method
name). If \code{NULL}, CI method's default settings are applied.}

\item{\code{autoEstimateCI}}{Logical. If \code{TRUE}, confidence intervals are
automatically estimated after optimization. If \code{FALSE}, the step is
skipped and can be triggered manually by calling the \code{estimateCI()}
method on the \code{ParameterIdentification} object.}

\item{\code{modelCostField}}{Read-only field name in the cost object used as the
optimization target. Currently, only \code{modelCost} is supported.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PIConfiguration-new}{\code{PIConfiguration$new()}}
\item \href{#method-PIConfiguration-print}{\code{PIConfiguration$print()}}
\item \href{#method-PIConfiguration-clone}{\code{PIConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-PIConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{PIConfiguration} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIConfiguration-print"></a>}}
\if{latex}{\out{\hypertarget{method-PIConfiguration-print}{}}}
\subsection{Method \code{print()}}{
Prints a summary of the \code{PIConfiguration}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PIConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
