% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIConfiguration.R
\docType{class}
\name{PIConfiguration}
\alias{PIConfiguration}
\title{PIConfiguration}
\description{
An object storing configuration for parameter identification
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{PIConfiguration}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{simulateSteadyState}}{Boolean for simulation steady-state first}

\item{\code{steadyStateTime}}{Time in minutes for steady-state simulation. May be NULL}

\item{\code{printEvaluationFeedback}}{Boolean. If \code{TRUE}, prints objective
function value after each evaluation. Default is \code{FALSE}}

\item{\code{simulationRunOptions}}{Object of \code{SimulationRunOptions} for simulation
runs. If \code{NULL}, default options are used.}

\item{\code{objectiveFunctionOptions}}{Configures model fit assessment
options. This setting influences how error and model fit are calculated
Supported options and their allowable values are detailed in
\code{ospsuite.parameteridentification::ObjectiveFunctionOptions}.}

\item{\code{algorithm}}{a string describing the optimization algorithm.
Supported algorithms are listed  in
\code{ospsuite.parameteridentification::Algorithms}.}

\item{\code{algorithmOptions}}{Named parameters for algorithm-specific options.
Supported options are listed  in \code{ospsuite.parameteridentification::AlgorithmOptions}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PIConfiguration-new}{\code{PIConfiguration$new()}}
\item \href{#method-PIConfiguration-print}{\code{PIConfiguration$print()}}
\item \href{#method-PIConfiguration-clone}{\code{PIConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-PIConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{PIConfiguration} object.
Print
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIConfiguration-print"></a>}}
\if{latex}{\out{\hypertarget{method-PIConfiguration-print}{}}}
\subsection{Method \code{print()}}{
prints a summary of the PIConfiguration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PIConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
