% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIConfiguration.R
\docType{class}
\name{PIConfiguration}
\alias{PIConfiguration}
\title{PIConfiguration}
\description{
Encapsulates configurations such as optimization algorithm choice,
and evaluation settings for parameter identification.
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{PIConfiguration}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{printEvaluationFeedback}}{Boolean. If \code{TRUE}, prints objective
function value after each evaluation. Default is \code{FALSE}.}

\item{\code{simulationRunOptions}}{Object of \code{SimulationRunOptions} for simulation
runs. If \code{NULL}, default options are used.}

\item{\code{objectiveFunctionOptions}}{Configures model fit evaluation settings,
influencing error and fit metrics. For option details and impact on cost metrics,
see \code{\link[ospsuite.parameteridentification]{ObjectiveFunctionSpecs}} and
\code{\link[ospsuite.parameteridentification]{calculateCostMetrics}}. Defaults found in
\code{\link[ospsuite.parameteridentification]{ObjectiveFunctionOptions}}.}

\item{\code{algorithm}}{A string specifying the optimization algorithm to use. See
\code{\link[ospsuite.parameteridentification]{Algorithms}} for a list of supported
algorithms. Defaults to \code{BOBYQA}.}

\item{\code{algorithmOptions}}{A list of named parameters for algorithm-specific
settings. Refer to \code{\link[ospsuite.parameteridentification]{AlgorithmOptions}}
for default settings per algorithm (e.g., \code{AlgorithmOptions_XYZ} where \code{XYZ}
denotes the algorithm name). If \code{NULL}, the algorithm's default settings
are applied.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PIConfiguration-new}{\code{PIConfiguration$new()}}
\item \href{#method-PIConfiguration-print}{\code{PIConfiguration$print()}}
\item \href{#method-PIConfiguration-clone}{\code{PIConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-PIConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{PIConfiguration} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIConfiguration-print"></a>}}
\if{latex}{\out{\hypertarget{method-PIConfiguration-print}{}}}
\subsection{Method \code{print()}}{
Prints a summary of the PIConfiguration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PIConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
