% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIOutputMapping.R
\docType{class}
\name{PIOutputMapping}
\alias{PIOutputMapping}
\title{PIOutputMapping}
\description{
An object that links together quantities from the simulation
to observed data. This object is passed to the ParameterIdentification
objects
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{PIOutputMapping}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{observedDataSets}}{Named list of \code{DataSet} objects that will be compared
to simulation results.}

\item{\code{dataTransformations}}{a named list of factors and offsets}

\item{\code{quantity}}{Simulation quantities which values are matched to the
observed data}

\item{\code{scaling}}{Linear (default) or logarithmic scaling for this output mapping}

\item{\code{transformResultsFunction}}{Function that will be applied to
simulated results. Allows to manipulate simulated values before
calculating the residuals. The function should manipulate numeric
vectors 'xVals' and 'yVals' (that being the simulated time- and
observation values) that will be than assigned to the x- and y-values
of the simulated result. The function must return a named list with key
'xVals' and 'yVals'.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PIOutputMapping-new}{\code{PIOutputMapping$new()}}
\item \href{#method-PIOutputMapping-addObservedDataSets}{\code{PIOutputMapping$addObservedDataSets()}}
\item \href{#method-PIOutputMapping-removeObservedDataSet}{\code{PIOutputMapping$removeObservedDataSet()}}
\item \href{#method-PIOutputMapping-setDataTransformations}{\code{PIOutputMapping$setDataTransformations()}}
\item \href{#method-PIOutputMapping-print}{\code{PIOutputMapping$print()}}
\item \href{#method-PIOutputMapping-clone}{\code{PIOutputMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIOutputMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-PIOutputMapping-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$new(quantity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantity}}{An object of the type \code{Quantity}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PIOutputMapping} object.
Add observed data as \code{DataSet} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIOutputMapping-addObservedDataSets"></a>}}
\if{latex}{\out{\hypertarget{method-PIOutputMapping-addObservedDataSets}{}}}
\subsection{Method \code{addObservedDataSets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$addObservedDataSets(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Object or a list of objects of the type
\code{DataSet}. Each data set must be of the same dimension as the simulation
quantity of the mapping.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If an observed data object with the same label already exists,
it will be overwritten.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIOutputMapping-removeObservedDataSet"></a>}}
\if{latex}{\out{\hypertarget{method-PIOutputMapping-removeObservedDataSet}{}}}
\subsection{Method \code{removeObservedDataSet()}}{
Remove the observed data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$removeObservedDataSet(label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{label of the x-y values series to be removed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIOutputMapping-setDataTransformations"></a>}}
\if{latex}{\out{\hypertarget{method-PIOutputMapping-setDataTransformations}{}}}
\subsection{Method \code{setDataTransformations()}}{
Set the data transformations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$setDataTransformations(
  labels = NULL,
  xOffsets = 0,
  yOffsets = 0,
  xFactors = 1,
  yFactors = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labels}}{A list of labels, each corresponding to one of the datasets. If labels are not specified, data transformations are set across all datasets}

\item{\code{xOffsets}}{A numeric list or a value of X-offsets}

\item{\code{yOffsets}}{A numeric list or a value of Y-offsets}

\item{\code{xFactors}}{A numeric list or a value of X-factors}

\item{\code{yFactors}}{A numeric list or a value of Y-factors}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIOutputMapping-print"></a>}}
\if{latex}{\out{\hypertarget{method-PIOutputMapping-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIOutputMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PIOutputMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
