% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIOutputMapping.R
\docType{class}
\name{PIOutputMapping}
\alias{PIOutputMapping}
\title{PIOutputMapping}
\description{
An object that combines simulation output with observed data. The parameter identification minimizes the
distance between the simulation output and the observed data.
}
\section{Super class}{
\code{\link[ospsuite:Printable]{ospsuite::Printable}} -> \code{PIOutputMapping}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{observedXYData}}{Named list with the \code{XYData} that will be
compared with simulation results. Names are the labels of the \code{xySeries} objects}

\item{\code{quantity}}{Simulation quantity which results are to be compared to observed data. Read-only.}

\item{\code{transformResultsFunction}}{Function that will be applied to simulated results. Allows to manipulate simulated values before
calculating the residuals. The function should manipulate numeric vectors 'xVals' and 'yVals' (that being the simulated time- and observation values)
that will be than assigned to the x- and y-values of the simulated result.
The function must return a named list with key 'xVals' and 'yVals'.
# @field residualsScaling Defines if the residuals are calculated based on the absolute distance (Simulation value - Observed value, scaling = "lin")
or as log(Simulation value) - log(Observed value) (scaling = "log"). Default is "lin".}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PIOutputMapping$new()}}
\item \href{#method-finalize}{\code{PIOutputMapping$finalize()}}
\item \href{#method-addObservedData}{\code{PIOutputMapping$addObservedData()}}
\item \href{#method-removeXYSeries}{\code{PIOutputMapping$removeXYSeries()}}
\item \href{#method-setXFactors}{\code{PIOutputMapping$setXFactors()}}
\item \href{#method-setYFactors}{\code{PIOutputMapping$setYFactors()}}
\item \href{#method-setXOffset}{\code{PIOutputMapping$setXOffset()}}
\item \href{#method-setConfiguration}{\code{PIOutputMapping$setConfiguration()}}
\item \href{#method-print}{\code{PIOutputMapping$print()}}
\item \href{#method-clone}{\code{PIOutputMapping$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$new(quantity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantity}}{An object of the type \code{Quantity}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PIOutputMapping` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clean up upon object removal
Add observed data as \code{OSPSTimeValues} object(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addObservedData"></a>}}
\if{latex}{\out{\hypertarget{method-addObservedData}{}}}
\subsection{Method \code{addObservedData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$addObservedData(OSPSTimeValues)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OSPSTimeValues}}{Object or a list of objects of the type \code{OSPSTimeValues}. Each data set must
be of the same dimension as the simulation quantity of the mapping.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If an observed data object with the same label already exists, it will be overwritten.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeXYSeries"></a>}}
\if{latex}{\out{\hypertarget{method-removeXYSeries}{}}}
\subsection{Method \code{removeXYSeries()}}{
Remove the observed data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$removeXYSeries(label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{label of the x-y values series to be removed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setXFactors"></a>}}
\if{latex}{\out{\hypertarget{method-setXFactors}{}}}
\subsection{Method \code{setXFactors()}}{
Set the X-factors of x-y values by labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$setXFactors(labels, xFactors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labels}}{A list of label of \code{XYData}}

\item{\code{xFactors}}{Numeric values that will be multiplied by the x-values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setYFactors"></a>}}
\if{latex}{\out{\hypertarget{method-setYFactors}{}}}
\subsection{Method \code{setYFactors()}}{
Set the y-factors of x-y values by labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$setYFactors(labels, yFactors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labels}}{A list of label of \code{XYData}}

\item{\code{yFactors}}{Numeric values that will be multiplied by the y-values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setXOffset"></a>}}
\if{latex}{\out{\hypertarget{method-setXOffset}{}}}
\subsection{Method \code{setXOffset()}}{
Set the X-offset of x-y values by labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$setXOffset(labels, xOffset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labels}}{A list of label of \code{XYData}}

\item{\code{xOffset}}{Numeric values that will be added to the x-values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setConfiguration"></a>}}
\if{latex}{\out{\hypertarget{method-setConfiguration}{}}}
\subsection{Method \code{setConfiguration()}}{
Apply settings stored in a dataMappingConfiguration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$setConfiguration(dataMappingConfiguration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataMappingConfiguration}}{An object of type \code{DataMappingConfiguration}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIOutputMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
