% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIParameters.R
\docType{class}
\name{PIParameters}
\alias{PIParameters}
\title{PIParameters}
\description{
A parameter to be optimized in a parameter identification routine
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{PIParameters}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parameters}}{List of parameter object. Adding or removing parameters is currently not supported.}

\item{\code{currValue}}{Current value of the parameters as used in the
simulation. The unit of the value is \code{PIParameters$unit}.}

\item{\code{startValue}}{Start value in the optimization.}

\item{\code{minValue}}{Minimal allowed value of the parameter.}

\item{\code{maxValue}}{Maximal allowed value of the parameter.}

\item{\code{unit}}{Unit of parameter values. If the unit is changed, the min/max/start values are NOT
adjusted automatically.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PIParameters-new}{\code{PIParameters$new()}}
\item \href{#method-PIParameters-setValue}{\code{PIParameters$setValue()}}
\item \href{#method-PIParameters-print}{\code{PIParameters$print()}}
\item \href{#method-PIParameters-clone}{\code{PIParameters$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-PIParameters-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
The start value is set to the value of the first parameter at time of creation
and can be changed with \code{PIParameters$startValue}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIParameters$new(parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{A list of objects of class \code{Parameter}.
All parameters will be optimized using the same value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PIParameters} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIParameters-setValue"></a>}}
\if{latex}{\out{\hypertarget{method-PIParameters-setValue}{}}}
\subsection{Method \code{setValue()}}{
Change the value of the parameter(s)
The unit of the value is $unit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIParameters$setValue(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Numerical value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-PIParameters-print}{}}}
\subsection{Method \code{print()}}{
Print the object to the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIParameters$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Rest arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PIParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PIParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PIParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
