% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{residualWeightingOptions}
\alias{residualWeightingOptions}
\title{Residual Weighting Methods for Cost Function}
\format{
An object of class \code{list} of length 2.
}
\usage{
residualWeightingOptions
}
\description{
Methods for weighting residuals in the model cost function, affecting how
discrepancies between simulations and observed data are evaluated.
}
\details{
The methods include:
\itemize{
\item \strong{\code{none}} – No weighting applied, treating all residuals equally.
\item \strong{\code{error}} – Weights based on error estimates for the dependent variable
in observed data.
}
}
\keyword{datasets}
