% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{residualWeightingOptions}
\alias{residualWeightingOptions}
\title{residualWeightingOptions
List of residual weighting methods used in the cost function.}
\format{
An object of class \code{list} of length 4.
}
\usage{
residualWeightingOptions
}
\description{
residualWeightingOptions
List of residual weighting methods used in the cost function.
}
\keyword{datasets}
