% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{robustMethodOptions}
\alias{robustMethodOptions}
\title{robustMethodOptions
List of robust weighting methods used in the cost function for handling outliers.}
\format{
An object of class \code{list} of length 3.
}
\usage{
robustMethodOptions
}
\description{
robustMethodOptions
List of robust weighting methods used in the cost function for handling outliers.
}
\keyword{datasets}
