% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ScalingOptions}
\alias{ScalingOptions}
\title{ScalingOptions
List of scaling options for output mappings.}
\format{
An object of class \code{list} of length 2.
}
\usage{
ScalingOptions
}
\description{
ScalingOptions
List of scaling options for output mappings.
}
\keyword{datasets}
