% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ScalingOptions}
\alias{ScalingOptions}
\title{Scaling Options for Output Mapping}
\format{
An object of class \code{list} of length 2.
}
\usage{
ScalingOptions
}
\description{
Scaling options applicable to output mappings, impacting how simulation
results are compared to observed data.
}
\details{
Available scaling options are:
\describe{
\item{\code{"lin"}}{Linear scaling, used by default.}
\item{\code{"log"}}{Logarithmic scaling, for when data spans several orders
of magnitude.}
}
}
\keyword{datasets}
